Rename files with _
