+++ src/pages/api/crmToken.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next'
-import GetCrmCloudToken from "../../usecases/api/getCrmCloudToken";
+import GetCrmCloudToken from "../../usecases/api/getCrmCloudToken_";
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
   const getCrmCloudToken = new GetCrmCloudToken();
